refactor RAG, move rag related data_pipeline, ollama model validation, and RAG to dedicated rag module
